Split app settings into config modules
